displayed actual value from topAdsGetFreeDeposit gql as expiry date in topads credit section adjusted ui a bit

+++ features/top_ads/top_ads_dashboard/src/main/java/com/tokopedia/topads/dashboard/data/model/ExpiryDateResponse.kt
@@ -16,7 +16,7 @@ data class ExpiryDateResponse(
             val depositID: Int = 0,
 
             @field:SerializedName("nominal")
-            val nominal: Int = 0,
+            val nominal: Float = 0f,
 
             @field:SerializedName("status")
             val status: Int = 0
